+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/attachments/content/FileAttachmentContent.kt
@@ -44,7 +44,6 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
-import coil.compose.rememberImagePainter
 import com.getstream.sdk.chat.utils.MediaStringUtil
 import io.getstream.chat.android.client.models.Attachment
 import io.getstream.chat.android.compose.R
@@ -52,6 +51,8 @@ import io.getstream.chat.android.compose.handlers.DownloadPermissionHandler.Comp
 import io.getstream.chat.android.compose.state.messages.attachments.AttachmentState
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
 import io.getstream.chat.android.compose.ui.util.MimeTypeIconProvider
+import io.getstream.chat.android.compose.ui.util.rememberStreamImagePainter
+import io.getstream.chat.android.offline.extensions.downloadAttachment
 
 /**
  * Builds a file attachment message which shows a list of files.
@@ -202,7 +203,7 @@ public fun FileAttachmentImage(attachment: Attachment) {
     val painter = if (isImage) {
         val dataToLoad = attachment.imageUrl ?: attachment.upload
 
-        rememberImagePainter(dataToLoad)
+        rememberStreamImagePainter(dataToLoad)
     } else {
         painterResource(id = MimeTypeIconProvider.getIconRes(attachment.mimeType))
     }